Update HtmlCode(1stPage).html
Moved things from the footer to the header like asked to. Added two paragraphs and an un-ordered list to #ARTICLE. Added things to both the header and the footer. Please read the comment under "<a href="">About Us</a>" and tell me if you think we can make that happen.